% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{addMetrics2colData}
\alias{addMetrics2colData}
\title{Add metrics to colData information}
\usage{
addMetrics2colData(se)
}
\arguments{
\item{se}{SummarizedExperiment object.}
}
\description{
This function adds the metrics data to the colData information from a
SummarizedExperiment object. This helps the subsetting of
SummarizedExperiment object while keeping the original values.
}
\examples{
data(se)
addMetrics2colData(se)
}
