% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{bcbcoldata}
\alias{bcbcoldata}
\title{Change colData information}
\usage{
bcbcoldata(se, sampleMetadata)
}
\arguments{
\item{se}{SummarizedExperiment object.}
}
\description{
This function will change the colData information from a
SummarizedExperiment object.
}
\examples{
data(se)
bcbcoldata(se, colData(se))
}
