% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio-Lite-package.R
\docType{package}
\name{bcbioLite-package}
\alias{bcbioLite}
\alias{bcbioLite-package}
\title{bcbioLite}
\description{
bcbioLite
}
\author{
\strong{Maintainer}: Victor Barrera Burgos \email{vbarrera@hpsh.harvard.edu}

}
