% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCBIO.R
\name{bcbreader}
\alias{bcbreader}
\title{bcbreader}
\usage{
bcbreader(projectDir)
}
\arguments{
\item{projectDir}{path to final bcbio project folder. Normally
named with date format.}
}
\description{
This function reads bcbio output.
}
\examples{
bcbreader()
}
\keyword{dataLoading}
