% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{bcbrun}
\alias{bcbrun}
\title{Normalize SummarizedExperiment from bcbreader function.}
\usage{
bcbrun(se, design = ~1)
}
\arguments{
\item{se}{SummarizedExperiment object generated by readBCBIO.}
}
\description{
This function will detect if the object contains the slots from
salmon quantification or featureCounts and normalize the data
using DESeq2. It will update raw, normalized and vst slots in a
SummarizedExperiment object.
}
\examples{
data(se)
bcbrun(se, design = ~condition)
}
