% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio-Lite-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{colData}
\alias{rowData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{rowData}}}
}}

