% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio-Lite-package.R
\docType{package}
\name{bcbioLite-package}
\alias{bcbioLite}
\alias{bcbioLite-package}
\title{bcbioLite}
\description{
Load results from BCBIO.
}
\author{
\strong{Maintainer}: Victor Barrera \email{Barrera.Vic@gmail.com} (\href{https://orcid.org/0000-0003-0590-4634}{ORCID})

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (\href{https://orcid.org/0000-0002-3859-3249}{ORCID})
}

}
