% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio-Lite-package.R
\docType{package}
\name{bcbioLite-package}
\alias{bcbioLite}
\alias{bcbioLite-package}
\title{bcbioLite}
\description{
bcbioLite
}
\author{
\strong{Maintainer}: Victor Barrera \email{Barrera.Vic@gmail.com} (0000-0003-0590-4634)

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (0000-0002-3859-3249)
}

}
